pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apartment-finder"
version = "0.1.0"
description = "Daily apartment finder that aggregates listings from multiple cities"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Emmanuel Chimezie"}
]
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pycraigslist>=0.7.0",
    "requests>=2.31",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
]
sendgrid = [
    "sendgrid>=6.10",
]

[project.scripts]
apartment-finder = "apartment_finder.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
